Move createProxy and rename config.ts
